backend/apps/notifications/serializers.py
"""
Notification serializers for API endpoints.
"""

from rest_framework import serializers
from .models import Notification


class NotificationSerializer(serializers.ModelSerializer):
    """Serializer for Notification model."""
    
    class Meta:
        model = Notification
        fields = ['id', 'type', 'title', 'message', 'link', 'data',
                  'is_read', 'read_at', 'created_at']
        read_only_fields = ['id', 'type', 'title', 'message', 'link', 
                           'data', 'created_at']
