backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
mysqlclient==2.2.1
python-dotenv==1.0.0
Pillow==10.2.0
django-filter==23.5
gunicorn==21.2.0
flake8==7.0.0
pytest==7.4.4
pytest-django==4.7.0
coverage==7.4.0
